[DeepClone] Support internal final classes with __unserialize (e.g. MongoDB BSON types)
Internal extension classes that are final and have a C-level create_object
handler cannot be instantiated via ReflectionClass::newInstanceWithoutConstructor()
or via unserialize('O:...:0:{}') when they require non-empty state.

MongoDB BSON types (ObjectId, Binary, UTCDateTime, etc.) fall into this
category: their __unserialize() rejects an empty array. They do, however,
round-trip correctly via serialize/unserialize once their full state is
provided, exactly as PHP's own serialization protocol does.

Track such classes in $needsFullUnserialize instead of throwing
NotInstantiableException. During reconstruction, eagerly finalize them
before the properties loop (when their state carries no object-ref masks)
by building the full O: serialized form and calling unserialize(), which
internally invokes __unserialize() with the correct state.

The states loop retains a fallback path for the rare case where the state
itself contains object references (requiring deferred resolution).
